agent/edgemindassistant/pyproject.toml

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "edgemindassistant"
version = "0.1.0"
requires-python = ">=3.10"

dependencies = [
    "bedrock-agentcore >= 1.0.3",
    "chromadb >= 0.4.0",
    "fastmcp >= 2.0.0",
    "httpx >= 0.27.0",
    "mcp >= 1.19.0",
    "pymupdf >= 1.24.0",
    "python-dotenv >= 1.2.1",
    "pyyaml >= 6.0",
    "strands-agents >= 1.13.0",
    "strands-agents-tools >= 0.1.0",
]